pyproject.toml
[project]
name = "vaf-am-build-03-identity"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.27.0",
    "pydantic>=2.6.3",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
